Revert "Move basic types to `TyContext` (#52)"
This reverts commit 6a623c9d30576cff878c14c55c787d3e1d9239e6.
