+++ crates/crane/src/typer.rs
@@ -1,10 +1,8 @@
 mod error;
-mod ty_context;
 mod r#type;
 
 pub use error::*;
 pub use r#type::*;
-pub use ty_context::*;
 
 use std::collections::HashMap;
 use std::sync::Arc;
@@ -47,21 +45,41 @@ struct ModuleItems {
     pub unions: HashMap<Ident, TyUnionDecl>,
 }
 
-pub struct Typer<'ctx> {
-    ctx: &'ctx TyContext,
-
+pub struct Typer {
     modules: HashMap<TyPath, ModuleItems>,
     use_map: HashMap<TyPath, TyPath>,
     scopes: Vec<HashMap<TyPath, Arc<Type>>>,
+
+    // Types.
+    unit_ty: Arc<Type>,
+    string_ty: Arc<Type>,
+    uint64_ty: Arc<Type>,
 }
 
-impl<'ctx> Typer<'ctx> {
-    pub fn new(ctx: &'ctx TyContext) -> Self {
+impl Typer {
+    pub fn new() -> Self {
+        let unit_ty = Arc::new(Type::UserDefined {
+            module: SmolStr::new_inline("std::prelude"),
+            name: SmolStr::new_inline("()"),
+        });
+
+        let string_ty = Arc::new(Type::UserDefined {
+            module: SmolStr::new_inline("std::prelude"),
+            name: SmolStr::new_inline("String"),
+        });
+
+        let uint64_ty = Arc::new(Type::UserDefined {
+            module: SmolStr::new_inline("std::prelude"),
+            name: SmolStr::new_inline("Uint64"),
+        });
+
         Self {
-            ctx,
             modules: HashMap::new(),
             use_map: HashMap::new(),
             scopes: Vec::new(),
+            unit_ty,
+            string_ty,
+            uint64_ty,
         }
     }
 
@@ -278,10 +296,10 @@ impl<'ctx> Typer<'ctx> {
                     name: "value".into(),
                     span: DUMMY_SPAN
                 },
-                ty: self.ctx.string.clone(),
+                ty: self.string_ty.clone(),
                 span: DUMMY_SPAN
             }],
-            self.ctx.unit.clone(),
+            self.unit_ty.clone(),
         )?;
         self.register_function(
             std_io_path,
@@ -294,10 +312,10 @@ impl<'ctx> Typer<'ctx> {
                     name: "value".into(),
                     span: DUMMY_SPAN
                 },
-                ty: self.ctx.string.clone(),
+                ty: self.string_ty.clone(),
                 span: DUMMY_SPAN
             }],
-            self.ctx.unit.clone(),
+            self.unit_ty.clone(),
         )?;
         self.register_function(
             std_int_path.clone(),
@@ -311,7 +329,7 @@ impl<'ctx> Typer<'ctx> {
                         name: "a".into(),
                         span: DUMMY_SPAN
                     },
-                    ty: self.ctx.uint64.clone(),
+                    ty: self.uint64_ty.clone(),
                     span: DUMMY_SPAN
                 },
                 TyFnParam {
@@ -319,11 +337,11 @@ impl<'ctx> Typer<'ctx> {
                         name: "b".into(),
                         span: DUMMY_SPAN
                     },
-                    ty: self.ctx.uint64.clone(),
+                    ty: self.uint64_ty.clone(),
                     span: DUMMY_SPAN
                 }
             ],
-            self.ctx.uint64.clone(),
+            self.uint64_ty.clone(),
         )?;
         self.register_function(
             std_int_path,
@@ -336,10 +354,10 @@ impl<'ctx> Typer<'ctx> {
                     name: "value".into(),
                     span: DUMMY_SPAN
                 },
-                ty: self.ctx.uint64.clone(),
+                ty: self.uint64_ty.clone(),
                 span: DUMMY_SPAN
             }],
-            self.ctx.string.clone(),
+            self.string_ty.clone(),
         )?;
 
         Ok(())
@@ -648,7 +666,7 @@ impl<'ctx> Typer<'ctx> {
         let params = self.infer_function_params(&function_decl.params)?;
 
         let return_ty = match function_decl.return_ty {
-            FnReturnTy::Unit => self.ctx.unit.clone(),
+            FnReturnTy::Unit => self.unit_ty.clone(),
             FnReturnTy::Ty(ref ty) => self.infer_ty(*ty.clone())?,
         };
 
@@ -961,7 +979,7 @@ impl<'ctx> Typer<'ctx> {
                 span,
             }),
             span,
-            ty: self.ctx.string.clone(),
+            ty: self.string_ty.clone(),
         })
     }
 
@@ -974,7 +992,7 @@ impl<'ctx> Typer<'ctx> {
                 span,
             }),
             span,
-            ty: self.ctx.uint64.clone(),
+            ty: self.uint64_ty.clone(),
         })
     }
 }
@@ -1002,9 +1020,7 @@ mod tests {
                 modules: thin_vec![module],
             };
 
-            let ty_context = TyContext::new();
-
-            let mut typer = Typer::new(&ty_context);
+            let mut typer = Typer::new();
 
             insta::assert_yaml_snapshot!(typer.type_check_package(package));
         })
